Added agent

+++ src/Models/AIResult.cs
@@ -1,3 +1,3 @@
 ﻿namespace BriefingTool.Models;
 
-public record AIResult(string? Output, string Debug, int TotalTokens);
+public record AIResult(string? Output, string Debug, long TotalTokens);
